requirements.txt
python-dotenv
torch
torchaudio
whisperx
pyannote.audio
yt-dlp
moviepy
numpy
soundfile
